Unnecessary cast
git-svn-id: https://svn.apache.org/repos/asf/commons/proper/compress/trunk@1308086 13f79535-47bb-0310-9956-ffa450edef68

+++ src/test/java/org/apache/commons/compress/archivers/tar/TarArchiveOutputStreamTest.java
@@ -88,11 +88,10 @@ public class TarArchiveOutputStreamTest extends AbstractTestCase {
         tos.write(new byte[10 * 1024]);
         byte[] data = bos.toByteArray();
         assertEquals(0x80,
-                     ((int) data[TarConstants.NAMELEN
-                                 + TarConstants.MODELEN
-                                 + TarConstants.UIDLEN
-                                 + TarConstants.GIDLEN]
-                      ) & 0x80);
+                     data[TarConstants.NAMELEN
+                        + TarConstants.MODELEN
+                        + TarConstants.UIDLEN
+                        + TarConstants.GIDLEN] & 0x80);
         TarArchiveInputStream tin =
             new TarArchiveInputStream(new ByteArrayInputStream(data));
         TarArchiveEntry e = tin.getNextTarEntry();
